♻️ refactor: Optimize Accountant storage and type safety
- Remove StorageMasks library and replace bit manipulation with explicit struct fields
- Enhance type safety by using uint128 for fees and amounts
- Simplify storage access and reduce complexity of fee and reward calculations
- Remove packed storage slot techniques in favor of more readable struct-based approach
- Update interfaces and tests to support new type-safe implementation

+++ packages/strategies/test/Base.t.sol
@@ -68,7 +68,9 @@ abstract contract BaseTest is Test {
     /*//////////////////////////////////////////////////////////////////////////
                                       HELPERS
     //////////////////////////////////////////////////////////////////////////*/
-    function _boundValidProtocolFee(uint256 newProtocolFee) internal view returns (uint256) {
-        return bound(newProtocolFee, 1, accountant.MAX_FEE_PERCENT() - accountant.getHarvestFeePercent());
+    function _boundValidProtocolFee(uint128 newProtocolFee) internal view returns (uint128) {
+        return
+            uint128(bound(uint256(newProtocolFee), 1, accountant.MAX_FEE_PERCENT() - accountant.getHarvestFeePercent()));
+            
     }
 }
